rediseño de index de supervision
